backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
anyio==3.7.1
watchfiles==0.21.0
email-validator>=2.0.0
